image annotation session refactor
